LayoutTests/http/tests/storageAccess/request-and-grant-access-with-per-page-scope-access-from-another-frame.html

<!DOCTYPE html>
<html>
<head>
    <script src="/js-test-resources/js-test.js"></script>
    <script src="/js-test-resources/ui-helper.js"></script>
    <script src="/resourceLoadStatistics/resources/util.js"></script>
</head>
<body>
    <script>
        description("Tests that storage access is granted for all frames under per-page storage scope.");
        jsTestIsAsync = true;

        const hostUnderTest = "localhost:8000";
        const statisticsUrl = "http://" + hostUnderTest + "/temp";

        const partitionHost = "127.0.0.1:8000";
        const thirdPartyOrigin = "http://localhost:8000";
        const resourcePath = "/storageAccess/resources";
        const thirdPartyBaseUrl = thirdPartyOrigin + resourcePath;
        const firstPartyCookieName = "firstPartyCookie";
        const subPathToSetFirstPartyCookie = "/set-cookie.php?name=" + firstPartyCookieName + "&value=value";
        const returnUrl = "http://" + partitionHost + "/storageAccess/request-and-grant-access-with-per-page-scope-access-from-another-frame.html";
        const subPathToGetCookies = "/get-cookies.php?name1=" + firstPartyCookieName;

        function openIframe(url, onLoadHandler) {
            const element = document.createElement("iframe");
            element.src = url;
            if (onLoadHandler) {
                element.onload = onLoadHandler;
            }
            document.body.appendChild(element);
        }

        function receiveMessage(event) {
            if (event.origin === "http://localhost:8000") {
                if (event.data.indexOf("PASS") !== -1)
                    testPassed(event.data.replace("PASS ", ""));
                else
                    testFailed(event.data);
            } else
                testFailed("Received a message from an unexpected origin: " + event.origin);
            runTest();
        }

        function activateElement(elementId) {
            var element = document.getElementById(elementId);
            var centerX = element.offsetLeft + element.offsetWidth / 2;
            var centerY = element.offsetTop + element.offsetHeight / 2;
            UIHelper.activateAt(centerX, centerY).then(
                function () {
                    if (window.eventSender)
                        eventSender.keyDown("escape");
                    else {
                        testFailed("No eventSender.");
                        setEnableFeature(false, finishJSTest);
                    }
                },
                function () {
                    testFailed("Promise rejected.");
                    setEnableFeature(false, finishJSTest);
                }
            );
        }

        function runTest() {
            switch (document.location.hash) {
                case "#step1":
                    // Set first-party cookie for localhost.
                    document.location.href = thirdPartyBaseUrl + subPathToSetFirstPartyCookie + "#" + returnUrl + "#step2";
                    break;
                case "#step2":
                    document.location.hash = "step3";
                    // Check that the first-party cookie does not get sent for localhost under 127.0.0.1.
                    openIframe(thirdPartyBaseUrl + subPathToGetCookies + "&message=Should not receive cookies without user interaction.", runTest);
                    break;
                case "#step3":
                    document.location.hash = "step4";
                    testRunner.setStatisticsHasHadUserInteraction(statisticsUrl, true, function() {
                        if (!testRunner.isStatisticsHasHadUserInteraction(statisticsUrl))
                            testFailed("Host did not get logged for user interaction.");
                        openIframe(thirdPartyBaseUrl + subPathToGetCookies + "&message=Should not receive cookies with user interaction.", runTest);
                    });
                    break;
                case "#step4":
                    document.location.hash = "step5";
                    let iframeElement = document.createElement("iframe");
                    iframeElement.onload = function() {
                        activateElement("TheIframeThatRequestsStorageAccess");
                    };
                    iframeElement.id = "TheIframeThatRequestsStorageAccess";
                    iframeElement.src = "http://localhost:8000/storageAccess/resources/request-storage-access-iframe.html#userShouldGrantAccess,userShouldBeConsulted,policyShouldGrantAccess,isNotSameOriginIframe";
                    document.body.appendChild(iframeElement);
                    break;
                case "#step5":
                    document.location.hash = "step6";
                    // Check that the cookie gets sent for localhost under 127.0.0.1 in a new frame.
                    openIframe(thirdPartyBaseUrl + subPathToGetCookies + "&message=Should receive cookies even though it's not the requesting frame.", runTest);
                    break;
                case "#step6":
                    testRunner.setStatisticsShouldBlockThirdPartyCookies(false, function() {
                        setEnableFeature(false, finishJSTest);
                    });
                    break;
            }
        }

        window.addEventListener("message", receiveMessage, false);

        if (document.location.hash === "") {
            setEnableFeature(true, function() {
                testRunner.dumpChildFramesAsText();
                document.location.hash = "step1";
                testRunner.setStatisticsShouldBlockThirdPartyCookies(true, runTest);
            });
        } else {
            runTest();
        }
    </script>
</body>
</html>
